refactor(Express): Move container and other to proper place
