+++ src/services/container.js
@@ -5,4 +5,20 @@ const getContainers = () => db.Container.findAll();
 
 const getContainerUsed = containerId => db.Rent.findOne({ where: { Container_ID: containerId } });
 
-export { getContainers, getContainerUsed };
+const getEmptyContainers = async () => {
+    const containers = await getContainers();
+
+    const emptyContainers = [];
+    for (const container of containers) {
+        const used = await getContainerUsed(container.ID);
+        if (!used) {
+            emptyContainers.push({
+                id: container.ID
+            });
+        }
+    }
+
+    return emptyContainers;
+};
+
+export { getContainers, getContainerUsed, getEmptyContainers };
