+++ src/services/rent.js
@@ -2,7 +2,6 @@
 import db from '../db/models';
 import { getPlant } from './plant';
 import { Op } from 'sequelize';
-import { getContainers, getContainerUsed } from './container';
 
 /* Get rent data from user GET users data */
 const getUserRentData = async userId => {
@@ -38,32 +37,6 @@ const getRentData = async rents => {
     return rentsData;
 };
 
-const newRent = async userId => {
-    const rent = await insertRent(userId);
-
-    // TODO: Auto assign container when avaliable
-    const emptyContainers = await getEmptyContainers();
-    if (emptyContainers.length !== 0) {
-        await assignContainer(rent.ID, emptyContainers[0].id);
-        // TODO: Send fill from email
-    }
-};
-
-const getEmptyContainers = async () => {
-    const containers = await getContainers();
-
-    const emptyContainers = [];
-    for (const container of containers) {
-        const used = await getContainerUsed(container.ID);
-        if (!used) {
-            emptyContainers.push({
-                id: container.ID
-            });
-        }
-    }
-
-    return emptyContainers;
-};
 
 const getUserRents = userId => db.Rent.findAll({ where: { User_ID: userId } });
 
@@ -85,4 +58,11 @@ const assignContainer = (ID, containerId, deadline = 5) =>
         Deadline: deadline
     }, { where: { ID } });
 
-export { getUserRentData, getUserRents, getOtherUserRentData, newRent };
+
+export {
+    getUserRentData,
+    getUserRents,
+    getOtherUserRentData,
+    insertRent,
+    assignContainer
+};
